+++ window_tools_fullset_comparer.py
@@ -4,9 +4,10 @@ import archive_games as games
 from threading import Thread
 
 
-class FileHashCheckerToolWindow():
-  def __init__(self, path: str):
+class FullsetComparerToolWindow():
+  def __init__(self, path: str, listbox_values: list):
     self.path = path
+    self.listbox_values = listbox_values
     self.abort = False
     self.thread: Thread = None
 
@@ -27,7 +28,7 @@ class FileHashCheckerToolWindow():
 
   def _create_window(self):
     window_layout = [
-      [common.create_modal_title("File checker tool")],
+      [common.create_modal_title("Fullset Comparer tool")],
       [sg.Multiline("", key="multiline_log", disabled=True, autoscroll=True, expand_x=True, expand_y=True)],
       [sg.Button("Close", expand_x=True, font=("", 10, "bold"), border_width=3, pad=(0, (10, 5)))]
     ]
@@ -40,20 +41,8 @@ class FileHashCheckerToolWindow():
     def add_line(text: str):
       self.window['multiline_log'].update(f"{text}\n", append=True)
     
-    import os, fnmatch, hashlib
-    format = list(games.games_dict.values())[0]['format']
+    import os
 
-    for file in fnmatch.filter(os.listdir(self.path), f"*.{format}"):
-      try:
-        with open(file, "rb") as stream: local_file_md5 = hashlib.md5(stream.read(-1)).hexdigest()
-        add_line("".join("-" for _ in range(100)))
-        add_line(f"Filename: {file}")
-        add_line(f"File Hash (md5): {local_file_md5}")
-        found = False
-        for game in games.games_dict:
-          if games.games_dict[game]['md5'] == local_file_md5:
-            add_line(f"Status: Found ({game})")
-            found = True
-            break
-        if not found: add_line(f"Status: Not found")
-      except OSError as error: add_line(f"{file} cannot be opened. ({error.strerror})")
+    local_game_list = ['{}.7z'.format(file[:file.rfind(".")]) for file in os.listdir(self.path)]
+    nointro_list = self.listbox_values
+    [add_line(n) for n in nointro_list if n not in local_game_list]